+++ src/main/java/com/example/examen1/service/impl/CodigoBarraImpl.java
@@ -25,16 +25,6 @@ public class CodigoBarraImpl implements CodigoBarraInt {
         return codigos;
     }
 
-    public ContentResponse<List<Producto>> consultaProductos(String codigo){
-        ContentResponse<List<Producto>> codigos = new ContentResponse<>();
-        List<Producto> listaCat = (List<Producto>) codigoBarraRepository.consultaProductosPorCodigo(codigo);
-        codigos.setData(listaCat);
-        codigos.setCodigo(1);
-        codigos.setError(false);
-        codigos.setStatus(StatusPeticion.EXITO);
-        return codigos;
-    }
-
     public ContentResponse<String> agregar(CodigoBarra codigoBarra){
         ContentResponse<String> response = new ContentResponse<>();
         try {
@@ -54,10 +44,16 @@ public class CodigoBarraImpl implements CodigoBarraInt {
     public ContentResponse<String> eliminar(Long idItem){
         ContentResponse<String> response = new ContentResponse<>();
         try {
-            codigoBarraRepository.deleteById(idItem);
             response.setCodigo(1);
-            response.setData("Eliminado");
             response.setStatus(StatusPeticion.EXITO);
+            if (codigoBarraRepository.existsById(idItem)){
+                codigoBarraRepository.deleteById(idItem);
+                response.setData("Eliminado");
+
+            }else{
+                response.setDescripcion("No se encontró el registro");
+            }
+
         }
         catch (Exception e){
             response.setDescripcion(e.getMessage());
@@ -71,13 +67,20 @@ public class CodigoBarraImpl implements CodigoBarraInt {
         ContentResponse<String> respuesta = new ContentResponse<>();
         respuesta.setStatus(StatusPeticion.ERROR);
         respuesta.setCodigo(0);
-        CodigoBarra codigoFinded = codigoBarraRepository.findById(codigoBarra.getId()).orElse(new CodigoBarra());
-        if (codigoFinded.getId()!= null){
-            CodigoBarra res = codigoBarraRepository.save(codigoBarra);
-            respuesta.setData(res.toString());
-            respuesta.setCodigo(1);
-            respuesta.setStatus(StatusPeticion.EXITO);
+        try {
+            if (codigoBarraRepository.existsById(codigoBarra.getId())){
+                CodigoBarra res = codigoBarraRepository.save(codigoBarra);
+                respuesta.setData(res.toString());
+                respuesta.setCodigo(1);
+                respuesta.setStatus(StatusPeticion.EXITO);
+            }else{
+                respuesta.setDescripcion("No se encontró el registro");
+            }
+
+        }catch (Exception e){
+            respuesta.setDescripcion(e.getMessage());
         }
+
         return respuesta;
     }
 }
